Refactor authentication functions
Resolves: #12
