+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-azure/compare/0.1.3...HEAD
 
+-   Refactoring: Moving authentication functions one module up
 -   Exporting discovery ability
 
 ## [0.1.3][] - 2018-06-06
